gruppper
bytte navnet til teammodal og createteam mot hverandre og la inn logikk slik at brukeren kun henter grupper brukeren er med i
