+++ hexacore/components/Team/CreateTeam.tsx
@@ -1,87 +1,149 @@
-import {v4 as uuid} from 'uuid';
-import React, { useState, useEffect} from 'react';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faFolderPlus, faSitemap, faUserCircle } from '@fortawesome/free-solid-svg-icons';
-import {Collapse, Input, Spacer, Text } from "@nextui-org/react";
-import { TestTeam } from './TestTeam';
-import { collection, getDocs, getDoc, addDoc, doc, query } from 'firebase/firestore';
-import { auth, db } from '../../firebase';
-import TeamModal from './TeamModal';
+import { app, auth, db, storage} from '../../firebase';
+import { doc, collection, addDoc, setDoc, getFirestore } from "firebase/firestore";
+import { Input } from '@nextui-org/react';
+import {useState, useEffect} from "react";
+import { v4 as uuidv4 } from 'uuid'
+import { ref, getDownloadURL, uploadBytesResumable } from "firebase/storage";
 
+const CreateTeam = ({isOpen, onClose}) => {
+    const userStorageRef = `users/groups`//id til gruppen skal her
+    const [name, setName] = useState("");
+    const [picture, setFileUrl] = useState(null)
 
-const CreateTeam = ()  => {
-
-  const [teams, setTeams] = React.useState([]);
+    const filechanged = async (e) =>{
+        var file = e.target.files[0];
+        const storageRef = ref(storage, `/groups/${auth.currentUser?.uid}`);
+        const uploadTask = uploadBytesResumable(storageRef, file);
   
+        uploadTask.on("state_changed",
+        (snapshot) => {
+        },
+        (error) => {
+          alert(error);
+        },
+        () => {
+          
+        getDownloadURL(uploadTask.snapshot.ref).then((downloadURL) => {
+            setFileUrl(downloadURL)
 
-  
-
-  useEffect(() => {
-    const fetchTeams = async () => {
-      const querySnapshot = await getDocs(query(collection(db, "groups")));
-      const fetchedTeams = querySnapshot.docs.map((doc) => {
-        const teamData = doc.data();
-        console.log(teamData.name)
-          return (
-            <Collapse.Group accordion={false}>
-            <Collapse title={
-              <button className="flex items-center w-full px-5 py-2 transition-colors duration-200 dark:hover:bg-gray-800 gap-x-2 hover:bg-gray-100 focus:outline-none">
-                {teamData.name}
-              </button>
-            }>
-              <div>
-                <div>
-                <img src={teamData.picture}/>
-                <Text className='flex items-center text-base font-normal text-gray-900 rounded-lg dark:text-white hover:bg-gray-100 dark:hover:bg-gray-700' >h</Text>
-                </div>
-                <Text className='flex items-center text-base font-normal text-gray-900 rounded-lg dark:text-white hover:bg-gray-100 dark:hover:bg-gray-700' >Table</Text>
-              </div>
-              </Collapse>                
-            </Collapse.Group>
-          );
-    });
-    
-    setTeams(fetchedTeams);
-  };
-  fetchTeams()
-    
-  }, []);
-
-
-
-
-  const [isModalOpen, setIsModalOpen] = useState(false);
-
-  function handleModalOpen() {
-    setIsModalOpen(true);
-  }
-
-  function handleModalClose() {
-    setIsModalOpen(false);
-  }
-
-
-
-  return (
-    <div>
-
-      <TeamModal isOpen={isModalOpen} onClose={handleModalClose} />
-
-  
-      
-       {teams.map((team) => (
-          <div>{team}</div>
-        ))}
-      <div className='grid place-items-center'>
-      <button type="button" onClick={handleModalOpen} className="text-white p-5 bg-[#24292F] hover:bg-[#24292F]/90 focus:ring-4 focus:outline-none focus:ring-[#24292F]/50 font-medium rounded-lg text-sm px-5 py-2.5 text-center inline-flex items-center dark:focus:ring-gray-500 dark:hover:bg-[#050708]/30 mr-2 mb-2">
-        <FontAwesomeIcon className='pr-2' icon={faFolderPlus}/>
+        });
+        }
+    );
+         
+   }
+    const submit = async () => {
+        onClose()
         
-        Create a team
-      </button>
-      </div>
+        const teamuid = uuidv4().replaceAll("-","");
 
-    </div>
-  )
-}
+        if (!name && !picture) {
+        console.log("none test")
+            return
+        } else {
+           await setDoc(doc(db, `groups/${teamuid}`), {
+                name: name,
+                ownerId: auth.currentUser?.uid,
+                teamuid: teamuid,
+                picture: picture
+            }).then(async()=>{
+              const user = auth.currentUser
+              setDoc(doc(db,`groups/${teamuid}/members/${teamuid}`), {
+                uid: user.uid,
+                name: user.displayName,
+                photo: user.photoURL
+              })
+              setDoc(doc(db, `users/${user.uid}/groups/${teamuid}/`),{
+                teamuid: teamuid
+              })
+            })
+            }
+        }
+        
+        
+
+    return (
+      <div
+        className={`${
+          isOpen ? 'block' : 'hidden'
+        } fixed z-10 inset-0 overflow-y-auto`}
+      >
+        <div className="flex items-center justify-center min-h-screen pt-4 px-4 pb-20 text-center sm:block sm:p-0">
+          <div
+            className="fixed inset-0 transition-opacity"
+            aria-hidden="true"
+            onClick={onClose}
+          >
+            <div className="absolute inset-0 bg-gray-500 opacity-75"></div>
+          </div>
+          <span
+            className="hidden sm:inline-block sm:align-middle sm:h-screen"
+            aria-hidden="true"
+          ></span>
+          <div
+            className={`${
+              isOpen ? 'block' : 'hidden'
+            } inline-block align-bottom bg-white rounded-lg px-4 pt-5 pb-4 text-left overflow-hidden shadow-xl transform transition-all sm:my-8 sm:align-middle sm:max-w-lg sm:w-full sm:p-6`}
+            role="dialog"
+            aria-modal="true"
+            aria-labelledby="modal-headline"
+          >
+            <div>
+              <div className="mt-3 text-center sm:mt-5">
+                <h3
+                  className="text-lg leading-6 font-medium text-gray-900"
+                  id="modal-headline"
+                >
+                  Create a team
+                </h3>
+                <div className="mt-2">
+                <div className="flex flex-col items-center pt-6 pr-6 pb-6 pl-6">
+
+                <Input
+                        name='emailRegisterInput'
+                        id='emailRegisterInput'
+                        type={"Name"}
+                        aria-label="Email input"
+                        onChange={e => { setName(e.currentTarget.value); }}
+                        clearable
+                        bordered            
+                        fullWidth
+                        color="primary"
+                        size="lg"
+                        placeholder="Team name"
+                    
+                    />
+                    <input
+                                        type="file"
+                                        className="appearance-none bg-gray-200 text-gray-900 px-2 py-1 shadow-sm border border-gray-400 rounded-md"
+                                        name="img"
+                                        onChange={filechanged}
+                                        required
+                                    />
+                <p className="mt-3 text-base leading-relaxed text-center text-black-200"></p>
+                
+                <div className="w-full mt-6">
+                  <a onClick={submit} className="flex text-center items-center justify-center w-full pt-4 pr-10 pb-4 pl-10 text-base
+                      font-medium text-white bg-indigo-600 rounded-xl transition duration-500 ease-in-out transform
+                      hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500">
+                      Create team</a>
+                </div>
+              </div>
+                </div>
+              </div>
+            </div>
+            <div className="mt-5 sm:mt-6">
+              <button
+                type="button"
+                className="inline-flex justify-center w-full rounded-md border border-gray-300 shadow-sm px-4 py-2 bg-white text-base font-medium text-gray-700 hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 sm:text-sm"
+                
+              >
+                Close
+              </button>
+            </div>
+          </div>
+        </div>
+      </div>
+    )
+  }
 
 export default CreateTeam
